# packages/api-client/spec/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "SpexCrafters API",
    "description": "REST API for the SpexCrafters Global B2B Optical Marketplace and Sourcing Platform. Sprint 1 surface: identity vertical slice (registration, email verification, login, token refresh, logout, current user). This committed spec is the contract of record; the Spring Boot application must produce a compatible document (validated in CI), and the TypeScript client in packages/api-client is generated from it.",
    "version": "0.1.0",
    "contact": { "name": "SpexCrafters Engineering" }
  },
  "servers": [
    { "url": "http://localhost:8080/api/v1", "description": "Local development" },
    { "url": "https://api.spexcrafters.com/api/v1", "description": "Production (working domain, pending OQ-8 confirmation)" }
  ],
...
  "paths": {
...
    "/suppliers/applications": {
      "post": {
        "tags": ["supplier"],
        "operationId": "createSupplierApplication",
        "summary": "Create a supplier identity, its DRAFT application and profile (requires supplier.create; one active supplier per organization)",
        "security": [{ "bearerAuth": [] }],
        "requestBody": { "required": true, "content": { "application/json": { "schema": { "$ref": "#/components/schemas/CreateSupplierApplicationRequest" } } } },
        "responses": {
          "201": { "description": "Application created.", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/SupplierApplication" } } } },
          "401": { "$ref": "#/components/responses/AuthenticationFailed" },
          "403": { "$ref": "#/components/responses/Forbidden" },
          "404": { "$ref": "#/components/responses/NotFound" },
          "409": { "$ref": "#/components/responses/Conflict" },
          "422": { "$ref": "#/components/responses/ValidationError" }
        }
      }
    },
    "/suppliers/applications/{applicationId}": {
      "parameters": [ { "name": "applicationId", "in": "path", "required": true, "schema": { "type": "string", "format": "uuid" } } ],
      "get": {
        "tags": ["supplier"],
        "operationId": "getSupplierApplication",
        "summary": "Get a supplier application (requires supplier.read)",
        "security": [{ "bearerAuth": [] }],
        "responses": {
          "200": { "description": "The application.", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/SupplierApplication" } } } },
          "401": { "$ref": "#/components/responses/AuthenticationFailed" },
          "404": { "$ref": "#/components/responses/NotFound" }
        }
      },
      "patch": {
        "tags": ["supplier"],
        "operationId": "updateSupplierApplicationDraft",
        "summary": "Update draft content (requires supplier.update; DRAFT/CHANGES_REQUESTED only; optimistic version)",
        "security": [{ "bearerAuth": [] }],
        "requestBody": { "required": true, "content": { "application/json": { "schema": { "$ref": "#/components/schemas/UpdateSupplierDraftRequest" } } } },
        "responses": {
          "200": { "description": "Updated.", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/SupplierApplication" } } } },
          "401": { "$ref": "#/components/responses/AuthenticationFailed" },
          "403": { "$ref": "#/components/responses/Forbidden" },
          "404": { "$ref": "#/components/responses/NotFound" },
          "409": { "$ref": "#/components/responses/Conflict" },
          "422": { "$ref": "#/components/responses/ValidationError" }
        }
      }
    },
    "/suppliers/applications/{applicationId}/submit": {
      "parameters": [ { "name": "applicationId", "in": "path", "required": true, "schema": { "type": "string", "format": "uuid" } } ],
      "post": {
        "tags": ["supplier"],
        "operationId": "submitSupplierApplication",
        "summary": "Submit (or resubmit) the application for review (requires supplier.submit; enforces completeness)",
        "security": [{ "bearerAuth": [] }],
        "responses": {
          "200": { "description": "Submitted.", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/SupplierApplication" } } } },
          "401": { "$ref": "#/components/responses/AuthenticationFailed" },
          "403": { "$ref": "#/components/responses/Forbidden" },
          "404": { "$ref": "#/components/responses/NotFound" },
          "409": { "$ref": "#/components/responses/Conflict" },
          "422": { "$ref": "#/components/responses/ValidationError" }
        }
      }
    },
    "/suppliers/applications/{applicationId}/withdraw": {
      "parameters": [ { "name": "applicationId", "in": "path", "required": true, "schema": { "type": "string", "format": "uuid" } } ],
      "post": {
        "tags": ["supplier"],
        "operationId": "withdrawSupplierApplication",
        "summary": "Withdraw the application (requires supplier.withdraw; frees the org's supplier slot)",
        "security": [{ "bearerAuth": [] }],
        "responses": {
          "200": { "description": "Withdrawn.", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/SupplierApplication" } } } },
          "401": { "$ref": "#/components/responses/AuthenticationFailed" },
          "403": { "$ref": "#/components/responses/Forbidden" },
          "404": { "$ref": "#/components/responses/NotFound" },
          "409": { "$ref": "#/components/responses/Conflict" }
        }
      }
    },
    "/suppliers/applications/{applicationId}/change-requests": {
      "parameters": [ { "name": "applicationId", "in": "path", "required": true, "schema": { "type": "string", "format": "uuid" } } ],
      "get": {
        "tags": ["supplier"],
        "operationId": "listChangeRequests",
        "summary": "List change requests raised against the application (requires supplier.read)",
        "security": [{ "bearerAuth": [] }],
        "responses": {
          "200": { "description": "Change requests.", "content": { "application/json": { "schema": { "type": "array", "items": { "$ref": "#/components/schemas/ReviewRequest" } } } } },
          "401": { "$ref": "#/components/responses/AuthenticationFailed" },
          "404": { "$ref": "#/components/responses/NotFound" }
        }
      }
    },
    "/suppliers/applications/{applicationId}/change-requests/{changeRequestId}/respond": {
      "parameters": [
        { "name": "applicationId", "in": "path", "required": true, "schema": { "type": "string", "format": "uuid" } },
        { "name": "changeRequestId", "in": "path", "required": true, "schema": { "type": "string", "format": "uuid" } }
      ],
      "post": {
        "tags": ["supplier"],
        "operationId": "respondToChangeRequest",
        "summary": "Respond to a change request (requires supplier.update)",
        "security": [{ "bearerAuth": [] }],
        "requestBody": { "required": true, "content": { "application/json": { "schema": { "$ref": "#/components/schemas/RespondChangeRequestRequest" } } } },
        "responses": {
          "200": { "description": "Response recorded.", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ReviewRequest" } } } },
          "401": { "$ref": "#/components/responses/AuthenticationFailed" },
          "403": { "$ref": "#/components/responses/Forbidden" },
          "404": { "$ref": "#/components/responses/NotFound" },
          "422": { "$ref": "#/components/responses/ValidationError" }
        }
      }
    },
...
    "/suppliers/{supplierId}/verification": {
      "parameters": [ { "name": "supplierId", "in": "path", "required": true, "schema": { "type": "string", "format": "uuid" } } ],
      "get": {
        "tags": ["verification"],
        "operationId": "getVerificationStatus",
        "summary": "Get scope-based verification status (requires supplier.verification.read)",
        "security": [{ "bearerAuth": [] }],
        "responses": {
          "200": { "description": "Verification status with per-scope results.", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/VerificationStatus" } } } },
          "401": { "$ref": "#/components/responses/AuthenticationFailed" },
          "404": { "$ref": "#/components/responses/NotFound" }
        }
      }
    },
...
    "/platform/review/suppliers/{applicationId}": {
      "parameters": [ { "name": "applicationId", "in": "path", "required": true, "schema": { "type": "string", "format": "uuid" } } ],
      "get": {
        "tags": ["platform"],
        "operationId": "getReviewDetail",
        "summary": "Reviewer detail: lifecycle, profile, evidence, change requests (requires supplier.review.read)",
        "security": [{ "bearerAuth": [] }],
        "responses": {
          "200": { "description": "The review detail.", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ReviewDetail" } } } },
          "401": { "$ref": "#/components/responses/AuthenticationFailed" },
          "403": { "$ref": "#/components/responses/Forbidden" },
          "404": { "$ref": "#/components/responses/NotFound" }
        }
      }
    },
...
    "/platform/suppliers/{supplierId}/verification/scopes/{scopeCode}/grant": {
      "parameters": [
        { "name": "supplierId", "in": "path", "required": true, "schema": { "type": "string", "format": "uuid" } },
        { "name": "scopeCode", "in": "path", "required": true, "schema": { "type": "string" } }
      ],
      "post": {
        "tags": ["platform"],
        "operationId": "grantVerificationScope",
        "summary": "Grant a verification scope with evidence linkage (requires supplier.verification.grant)",
        "security": [{ "bearerAuth": [] }],
        "requestBody": { "required": true, "content": { "application/json": { "schema": { "$ref": "#/components/schemas/GrantScopeRequest" } } } },
        "responses": {
          "200": { "description": "Granted.", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/VerificationStatus" } } } },
          "401": { "$ref": "#/components/responses/AuthenticationFailed" },
          "403": { "$ref": "#/components/responses/Forbidden" },
          "404": { "$ref": "#/components/responses/NotFound" },
          "422": { "$ref": "#/components/responses/ValidationError" }
        }
      }
    },
    "/platform/suppliers/{supplierId}/verification/scopes/{scopeCode}/suspend": {
      "parameters": [
        { "name": "supplierId", "in": "path", "required": true, "schema": { "type": "string", "format": "uuid" } },
        { "name": "scopeCode", "in": "path", "required": true, "schema": { "type": "string" } }
      ],
      "post": {
        "tags": ["platform"],
        "operationId": "suspendVerificationScope",
        "summary": "Suspend a granted verification scope (requires supplier.verification.suspend)",
        "security": [{ "bearerAuth": [] }],
        "requestBody": { "required": false, "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ReasonRequest" } } } },
        "responses": {
          "200": { "description": "Suspended.", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/VerificationStatus" } } } },
          "401": { "$ref": "#/components/responses/AuthenticationFailed" },
          "403": { "$ref": "#/components/responses/Forbidden" },
          "404": { "$ref": "#/components/responses/NotFound" }
        }
      }
    },
    "/platform/suppliers/{supplierId}/verification/scopes/{scopeCode}/revoke": {
      "parameters": [
        { "name": "supplierId", "in": "path", "required": true, "schema": { "type": "string", "format": "uuid" } },
        { "name": "scopeCode", "in": "path", "required": true, "schema": { "type": "string" } }
      ],
      "post": {
        "tags": ["platform"],
        "operationId": "revokeVerificationScope",
        "summary": "Revoke a verification scope (requires supplier.verification.revoke)",
        "security": [{ "bearerAuth": [] }],
        "requestBody": { "required": false, "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ReasonRequest" } } } },
        "responses": {
          "200": { "description": "Revoked.", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/VerificationStatus" } } } },
          "401": { "$ref": "#/components/responses/AuthenticationFailed" },
          "403": { "$ref": "#/components/responses/Forbidden" },
          "404": { "$ref": "#/components/responses/NotFound" }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": { "type": "http", "scheme": "bearer", "bearerFormat": "JWT" }
    },
    "schemas": {
...
      "SupplierApplicationStatus": {
        "type": "string",
        "enum": ["DRAFT", "SUBMITTED", "UNDER_REVIEW", "CHANGES_REQUESTED", "RESUBMITTED", "APPROVED", "REJECTED", "WITHDRAWN"],
        "description": "Clients must tolerate unknown future values."
      },
      "OperationalStatus": {
        "type": "string",
        "enum": ["PENDING", "ACTIVE", "SUSPENDED", "DEACTIVATED"]
      },
...
      "ScanStatus": {
        "type": "string",
        "enum": ["PENDING_SCAN", "SCANNING", "CLEAN", "REJECTED", "QUARANTINED"],
        "description": "Phase 7 never marks evidence CLEAN (deferred scanner)."
      },
      "EvidenceUploadState": {
        "type": "string",
        "enum": ["INITIATED", "FINALIZED"]
      },
      "EvidenceReviewStatus": {
        "type": "string",
        "enum": ["UNREVIEWED", "ACCEPTED", "REJECTED"]
      },
      "RetentionStatus": {
        "type": "string",
        "enum": ["NONE", "RETAINED"]
      },
      "VerificationScopeStatus": {
        "type": "string",
        "enum": ["NOT_REQUESTED", "PENDING", "UNDER_REVIEW", "VERIFIED", "REJECTED", "CHANGES_REQUESTED", "EXPIRED", "SUSPENDED", "REVOKED"],
        "description": "Scope-based; there is no verified boolean anywhere."
      },
...
      "SupplierCapability": {
        "type": "string",
        "enum": ["supplier.create", "supplier.read", "supplier.update", "supplier.submit", "supplier.withdraw", "supplier.evidence.read", "supplier.evidence.upload", "supplier.evidence.delete", "supplier.verification.read"],
        "description": "Org-scoped supplier capabilities; clients must tolerate unknown future values."
      },
      "CreateSupplierApplicationRequest": {
        "type": "object",
        "required": ["organizationId", "originalLocale", "legalName"],
        "properties": {
          "organizationId": { "type": "string", "format": "uuid" },
          "originalLocale": { "type": "string", "description": "BCP 47 code of the authoring language" },
          "legalName": { "type": "string", "maxLength": 300, "description": "Class E — never machine-translated" }
        }
      },
      "SupplierApplication": {
        "type": "object",
        "required": ["applicationId", "supplierId", "organizationId", "originalLocale", "status", "operationalStatus", "createdAt", "version", "callerCapabilities"],
        "properties": {
          "applicationId": { "type": "string", "format": "uuid" },
          "supplierId": { "type": "string", "format": "uuid" },
          "organizationId": { "type": "string", "format": "uuid" },
          "originalLocale": { "type": "string" },
          "status": { "$ref": "#/components/schemas/SupplierApplicationStatus" },
          "operationalStatus": { "$ref": "#/components/schemas/OperationalStatus" },
          "submittedAt": { "type": "string", "format": "date-time", "nullable": true },
          "decidedAt": { "type": "string", "format": "date-time", "nullable": true },
          "createdAt": { "type": "string", "format": "date-time" },
          "version": { "type": "integer", "description": "Optimistic-locking version for draft updates" },
          "callerCapabilities": { "type": "array", "items": { "$ref": "#/components/schemas/SupplierCapability" } }
        }
      },
      "UpdateSupplierDraftRequest": {
        "type": "object",
        "required": ["version"],
        "properties": {
          "legalName": { "type": "string", "maxLength": 300 },
          "registeredLegalNameTranslated": { "type": "string", "maxLength": 300 },
          "tradingName": { "type": "string", "maxLength": 300 },
          "registrationNumber": { "type": "string", "maxLength": 120 },
          "countryOfRegistration": { "type": "string", "minLength": 2, "maxLength": 2 },
          "registrationAuthority": { "type": "string", "maxLength": 300 },
          "registrationDate": { "type": "string", "format": "date" },
          "companyTypeCode": { "type": "string" },
          "yearEstablished": { "type": "integer" },
          "employeeRange": { "type": "string" },
          "website": { "type": "string" },
          "businessEmail": { "type": "string" },
          "businessPhone": { "type": "string" },
          "types": { "type": "array", "items": { "type": "string" }, "description": "Supplier type codes (replace-all)" },
          "capabilities": { "type": "array", "items": { "type": "string" }, "description": "Declared capability codes (replace-all)" },
          "version": { "type": "integer", "description": "Optimistic-locking version; mismatch → 409" }
        }
      },
      "RespondChangeRequestRequest": {
        "type": "object",
        "required": ["response"],
        "properties": {
          "response": { "type": "string", "maxLength": 4000 },
          "responseLocale": { "type": "string" }
        }
      },
      "ReviewRequest": {
        "type": "object",
        "required": ["id", "applicationId", "requestedItem", "reason", "status", "requestedAt"],
        "properties": {
          "id": { "type": "string", "format": "uuid" },
          "applicationId": { "type": "string", "format": "uuid" },
          "requestedItem": { "type": "string" },
          "reason": { "type": "string" },
          "status": { "type": "string", "enum": ["OPEN", "RESPONDED", "RESOLVED"] },
          "supplierResponse": { "type": "string", "nullable": true },
          "responseLocale": { "type": "string", "nullable": true },
          "requestedAt": { "type": "string", "format": "date-time" },
          "resolvedAt": { "type": "string", "format": "date-time", "nullable": true }
        }
      },
      "CapabilityDeclaration": {
        "type": "object",
        "required": ["capabilityCode", "claimStatus"],
        "properties": {
          "capabilityCode": { "type": "string" },
          "claimStatus": { "$ref": "#/components/schemas/ClaimStatus" }
        }
      },
      "ProfileTranslation": {
        "type": "object",
        "required": ["locale", "original", "translationStatus", "translationSource", "sourceLocale", "sourceVersion", "stale"],
        "properties": {
          "locale": { "type": "string" },
          "original": { "type": "boolean", "description": "The authoritative original-language row" },
          "translationStatus": { "$ref": "#/components/schemas/TranslationStatus" },
          "translationSource": { "$ref": "#/components/schemas/TranslationSource" },
          "sourceLocale": { "type": "string" },
          "sourceVersion": { "type": "integer" },
          "stale": { "type": "boolean", "description": "Source content advanced past this translation" },
          "tradingName": { "type": "string", "nullable": true },
          "companyDescription": { "type": "string", "nullable": true },
          "productionCapabilityDescription": { "type": "string", "nullable": true },
          "oemDescription": { "type": "string", "nullable": true },
          "odmDescription": { "type": "string", "nullable": true },
          "privateLabelDescription": { "type": "string", "nullable": true },
          "qualityControlDescription": { "type": "string", "nullable": true },
          "exportMarketDescription": { "type": "string", "nullable": true }
        }
      },
...
      "Facility": {
        "type": "object",
        "required": ["id", "facilityTypeCode", "country", "addressPrivacy", "ownership", "isPublic", "sourceVersion", "translations"],
        "properties": {
          "id": { "type": "string", "format": "uuid" },
          "facilityTypeCode": { "type": "string" },
          "country": { "type": "string" },
          "region": { "type": "string", "nullable": true },
          "city": { "type": "string", "nullable": true },
          "addressPrivacy": { "$ref": "#/components/schemas/AddressPrivacy" },
          "ownership": { "$ref": "#/components/schemas/FacilityOwnership" },
          "isPublic": { "type": "boolean" },
          "sourceVersion": { "type": "integer" },
          "translations": { "type": "array", "items": { "$ref": "#/components/schemas/FacilityTranslation" } }
        }
      },
      "SupplierProfile": {
        "type": "object",
        "required": ["supplierId", "originalLocale", "legalName", "sourceVersion", "types", "capabilities", "facilities", "translations"],
        "properties": {
          "supplierId": { "type": "string", "format": "uuid" },
          "originalLocale": { "type": "string" },
          "legalName": { "type": "string" },
          "registeredLegalNameTranslated": { "type": "string", "nullable": true },
          "tradingName": { "type": "string", "nullable": true },
          "registrationNumber": { "type": "string", "nullable": true },
          "countryOfRegistration": { "type": "string", "nullable": true },
          "registrationAuthority": { "type": "string", "nullable": true },
          "registrationDate": { "type": "string", "format": "date", "nullable": true },
          "companyTypeCode": { "type": "string", "nullable": true },
          "yearEstablished": { "type": "integer", "nullable": true },
          "employeeRange": { "type": "string", "nullable": true },
          "website": { "type": "string", "nullable": true },
          "businessEmail": { "type": "string", "nullable": true },
          "businessPhone": { "type": "string", "nullable": true },
          "sourceVersion": { "type": "integer" },
          "types": { "type": "array", "items": { "type": "string" } },
          "capabilities": { "type": "array", "items": { "$ref": "#/components/schemas/CapabilityDeclaration" } },
          "facilities": { "type": "array", "items": { "$ref": "#/components/schemas/Facility" } },
          "translations": { "type": "array", "items": { "$ref": "#/components/schemas/ProfileTranslation" } }
        }
      },
...
      "Evidence": {
        "type": "object",
        "required": ["id", "supplierId", "evidenceTypeCode", "originalFilename", "uploadState", "scanStatus", "reviewStatus", "retentionStatus", "downloadable"],
        "properties": {
          "id": { "type": "string", "format": "uuid" },
          "supplierId": { "type": "string", "format": "uuid" },
          "evidenceTypeCode": { "type": "string" },
          "originalFilename": { "type": "string" },
          "mediaType": { "type": "string", "nullable": true },
          "byteSize": { "type": "integer", "format": "int64", "nullable": true },
          "sha256": { "type": "string", "nullable": true },
          "documentLocale": { "type": "string", "nullable": true },
          "uploadState": { "$ref": "#/components/schemas/EvidenceUploadState" },
          "scanStatus": { "$ref": "#/components/schemas/ScanStatus" },
          "reviewStatus": { "$ref": "#/components/schemas/EvidenceReviewStatus" },
          "retentionStatus": { "$ref": "#/components/schemas/RetentionStatus" },
          "downloadable": { "type": "boolean", "description": "Fail-closed derived flag" },
          "uploadedAt": { "type": "string", "format": "date-time", "nullable": true }
        }
      },
      "VerificationScopeResult": {
        "type": "object",
        "required": ["scopeCode", "status", "evidenceIds"],
        "properties": {
          "scopeCode": { "type": "string" },
          "status": { "$ref": "#/components/schemas/VerificationScopeStatus" },
          "decidedAt": { "type": "string", "format": "date-time", "nullable": true },
          "validFrom": { "type": "string", "format": "date-time", "nullable": true },
          "validUntil": { "type": "string", "format": "date-time", "nullable": true },
          "reason": { "type": "string", "nullable": true },
          "evidenceIds": { "type": "array", "items": { "type": "string", "format": "uuid" } }
        }
      },
      "VerificationStatus": {
        "type": "object",
        "required": ["supplierId", "caseStatus", "scopes"],
        "properties": {
          "supplierId": { "type": "string", "format": "uuid" },
          "caseStatus": { "$ref": "#/components/schemas/VerificationScopeStatus" },
          "openedAt": { "type": "string", "format": "date-time", "nullable": true },
          "scopes": { "type": "array", "items": { "$ref": "#/components/schemas/VerificationScopeResult" } }
        }
      },
      "GrantScopeRequest": {
        "type": "object",
        "required": ["evidenceIds"],
        "properties": {
          "evidenceIds": { "type": "array", "minItems": 1, "items": { "type": "string", "format": "uuid" }, "description": "Evidence linkage is required for a grant" },
          "validUntil": { "type": "string", "format": "date-time" },
          "reason": { "type": "string", "maxLength": 4000 }
        }
      },
      "ReasonRequest": {
        "type": "object",
        "properties": { "reason": { "type": "string", "maxLength": 4000 } }
      },
...
      "ReviewDetail": {
        "type": "object",
        "required": ["applicationId", "supplierId", "organizationId", "status", "operationalStatus", "version", "profile", "evidence", "changeRequests"],
        "properties": {
          "applicationId": { "type": "string", "format": "uuid" },
          "supplierId": { "type": "string", "format": "uuid" },
          "organizationId": { "type": "string", "format": "uuid" },
          "status": { "$ref": "#/components/schemas/SupplierApplicationStatus" },
          "operationalStatus": { "$ref": "#/components/schemas/OperationalStatus" },
          "submittedAt": { "type": "string", "format": "date-time", "nullable": true },
          "version": { "type": "integer" },
          "profile": { "$ref": "#/components/schemas/SupplierProfile" },
          "evidence": { "type": "array", "items": { "$ref": "#/components/schemas/Evidence" } },
          "changeRequests": { "type": "array", "items": { "$ref": "#/components/schemas/ReviewRequest" } }
        }
      },
...
      "Problem": {
        "type": "object",
        "description": "RFC 9457 problem details. Content type application/problem+json.",
        "required": ["type", "title", "status"],
        "properties": {
          "type": { "type": "string", "format": "uri" },
          "title": { "type": "string" },
          "status": { "type": "integer" },
          "detail": { "type": "string" },
          "instance": { "type": "string" },
          "correlationId": { "type": "string" },
          "errors": {
            "type": "array",
            "items": {
              "type": "object",
              "required": ["field", "code", "message"],
              "properties": {
                "field": { "type": "string" },
                "code": { "type": "string" },
                "message": { "type": "string", "description": "Locale-resolved; code maps to frontend i18n keys" }
              }
            }
          }
        }
      }
    },
    "responses": {
      "ValidationError": {
        "description": "Request failed validation.",
        "content": { "application/problem+json": { "schema": { "$ref": "#/components/schemas/Problem" } } }
      },
      "AuthenticationFailed": {
        "description": "Missing/invalid credentials or token.",
        "content": { "application/problem+json": { "schema": { "$ref": "#/components/schemas/Problem" } } }
      },
...
      "Conflict": {
        "description": "Resource conflict (e.g. email already registered, optimistic-lock version mismatch, duplicate membership, last-owner invariant).",
        "content": { "application/problem+json": { "schema": { "$ref": "#/components/schemas/Problem" } } }
      },
      "Forbidden": {
        "description": "Authenticated member lacks the required capability (problem type .../problems/authorization).",
        "content": { "application/problem+json": { "schema": { "$ref": "#/components/schemas/Problem" } } }
      },
      "NotFound": {
        "description": "Resource absent — or concealed from non-members by tenancy policy.",
        "content": { "application/problem+json": { "schema": { "$ref": "#/components/schemas/Problem" } } }
      },
...
    }
  }
}